Date: 20180608
File: portage-latest.tar.xz
Verify result:
gpg: Signature made Fri Jun  8 02:51:24 2018 CEST
gpg:                using RSA key E1D6ABB63BFCFB4BA02FDF1CEC590EEAC9189250
gpg: Good signature from "Gentoo Portage Snapshot Signing Key (Automated Signing Key)" [ultimate]

gpgsig:
-----BEGIN PGP SIGNATURE-----

iQKTBAABCgB9FiEE4dartjv8+0ugL98c7FkO6skYklAFAlsZ0wxfFIAAAAAALgAo
aXNzdWVyLWZwckBub3RhdGlvbnMub3BlbnBncC5maWZ0aGhvcnNlbWFuLm5ldEUx
RDZBQkI2M0JGQ0ZCNEJBMDJGREYxQ0VDNTkwRUVBQzkxODkyNTAACgkQ7FkO6skY
klBuFRAAhA6YnWmuiuebSzn+tZzygBKQtNbMIZTbe0kUMnSQHQCMJpj6HBgdvOcK
bcEBnd1qIfZUBHXyOaT5I0UtDYZ9n7X/WWVKmA/Ki7/SlQqMF8Sa3J/IAGuvMAsj
guyrFoJy8u27aGTflU0u79qeLN+CyGLYuc3b6Ykoft5hUc6D7Do0Rw5l1ei7705y
lDzTKtaBM0jtU9ZeRcj8lp7JeZSR3G8meZzPff1sktcBLa1Yz+QvQPER5u14Zokc
fVcEYNcrb+xxdIRzReSkkWl4TxAFAKWub289qHbJ2CvFtRu+7DJXingYa57wAB3C
lg15juf6zL25KPphSp/AfJET12JuAUtjAUpTghxNlMeyMXr7ddirRwS6cU8xmoUi
4PC6xOaGixR0M1uLQB8uMd103S6sL7HQOIicaa3qhugfOMlfpvP37Uvsb9M4CAAK
gHpmZLQ4xzjk2XpBw/Ox2IO6CIKQXhw0k1TXwvItttV2f7+1ecYYpZ1GwEq2rhzN
YAhTxm3V1DvRE6EJY4zQ6kBv+BeZ3jTV+wTUwuAjU0h03xTFmGw9XFP/cHSPTZH9
XP27hhFGFPGLhgnoWxjMepnYx++9IzJMbnLD3fFVaoP2AxLl27bV148hNmdpTebd
+YiSxfOMxTjq3yQ9Nq0wCcOrhy/EOky7WPnmG160oOpxCy2ebfQ=
=pWjj
-----END PGP SIGNATURE-----

+++ media-libs/mesa/mesa-18.1.1-r1.ebuild
@@ -3,7 +3,7 @@
 
 EAPI=6
 
-EGIT_REPO_URI="https://anongit.freedesktop.org/git/mesa/mesa.git"
+EGIT_REPO_URI="https://gitlab.freedesktop.org/mesa/mesa.git"
 
 if [[ ${PV} = 9999 ]]; then
 	GIT_ECLASS="git-r3"
@@ -401,11 +401,13 @@ multilib_src_configure() {
 multilib_src_install() {
 	emake install DESTDIR="${D}"
 
-	# These files are now provided by >=dev-libs/wayland-1.15.0
-	rm "${ED}/usr/$(get_libdir)/libwayland-egl.so" || die
-	rm "${ED}/usr/$(get_libdir)/libwayland-egl.so.1" || die
-	rm "${ED}/usr/$(get_libdir)/libwayland-egl.so.1.0.0" || die
-	rm "${ED}/usr/$(get_libdir)/pkgconfig/wayland-egl.pc" || die
+	if use wayland; then
+		# These files are now provided by >=dev-libs/wayland-1.15.0
+		rm "${ED}/usr/$(get_libdir)/libwayland-egl.so" || die
+		rm "${ED}/usr/$(get_libdir)/libwayland-egl.so.1" || die
+		rm "${ED}/usr/$(get_libdir)/libwayland-egl.so.1.0.0" || die
+		rm "${ED}/usr/$(get_libdir)/pkgconfig/wayland-egl.pc" || die
+	fi
 
 	if use opencl; then
 		ebegin "Moving Gallium/Clover OpenCL implementation for dynamic switching"
